added suggested changes

+++ app/controllers/api/v1/garden_plants_controller.rb
@@ -16,7 +16,14 @@ class Api::V1::GardenPlantsController < ApplicationController
     end
 
     def destroy
-        garden_plant = GardenPlant.find_by(garden_id: params[:garden_id], plant_id: params[:plant_id])
+        garden = Garden.find_by(id: params[:garden_id])
+        
+        if garden.nil?
+          render json: { error: 'Garden not found' }, status: :not_found
+          return
+        end
+      
+        garden_plant = GardenPlant.find_by(garden_id: garden.id, plant_id: params[:plant_id])
       
         if garden_plant
           garden_plant.destroy
@@ -24,7 +31,7 @@ class Api::V1::GardenPlantsController < ApplicationController
         else
           render json: { error: 'Garden plant not found' }, status: :not_found
         end
-    end
+      end
 
     private
     def plant_params
